+++ scrapli_ssh2/__init__.py
@@ -1,3 +1,3 @@
 """ssh2-python transport plugin for scrapli"""
 
-__version__ = "2020.10.07"
+__version__ = "2020.10.10"